docs: Proposal doc prepared for Capstone Project
Proposal document has detail of dataset, feature, Data exploration and different model that will be tried
for best analyzing the data. It also contains information on how Overfitting will be avoided and performance will be  improved.
It has information on what will be benchmark model and how gradually we will improve to perform best.